rename: `FollowingsList`에서 `FollowingList`로 이름 변경
